backend/requirements.txt
Django>=4.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
gunicorn
